Add BulkUpdate::TraineeUploads::BasePolicy
