+++ app/policies/bulk_update/details/trainee_upload_policy.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module BulkUpdate
+  module Details
+    class TraineeUploadPolicy
+    end
+  end
+end
