+++ app/policies/bulk_update/trainee_upload_policy.rb
@@ -1,27 +1,7 @@
 # frozen_string_literal: true
 
 module BulkUpdate
-  class TraineeUploadPolicy
-    class Scope
-      attr_reader :user, :scope
-
-      def initialize(user, scope)
-        @user  = user
-        @scope = scope
-      end
-
-      def resolve
-        scope.where(provider: user.organisation)
-      end
-    end
-
-    attr_reader :user, :trainee_upload
-
-    def initialize(user, trainee_upload)
-      @user           = user
-      @trainee_upload = trainee_upload
-    end
-
+  class TraineeUploadPolicy < TraineeUploads::BasePolicy
     def new?
       user.hei_provider?
     end
